Chore: Separate model types with model directory
