+++ src/components/AdjacentPostCard.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Post } from '@/service/posts';
+import { Post } from '@/model/post';
 import Image from 'next/image';
 import Link from 'next/link';
 import { FiArrowLeftCircle, FiArrowRightCircle } from 'react-icons/fi';
