+++ src/components/Comments.tsx
@@ -6,21 +6,6 @@ type Props = {
   postId: string;
 };
 
-export type User = {
-  uid: string;
-  photoURL: string;
-  email: string;
-  displayName: string;
-};
-
-export type Comment = {
-  postId: string;
-  id: string;
-  content: string;
-  createdAt: string;
-  user: User;
-};
-
 export default function Comments({ postId }: Props) {
   return (
     <CommentsContextProvider postId={postId}>
