+++ src/components/PostCardImage.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { useState } from 'react';
 import Image from 'next/image';
-import { Post } from '@/service/posts';
+import { Post } from '@/model/post';
 
 type Props = {
   post: Post;
