+++ src/model/user.ts
@@ -1,0 +1,8 @@
+export type User = {
+  uid: string;
+  photoURL: string;
+  email: string;
+  displayName: string;
+};
+
+export type SimpleUser = Omit<User, 'email'>;
